28/08/2025 -#sửa nhanh qua giao diện bên phía người dùng(navbar, content,...), làm fe mêt vcl
