+++ src/components/layout/client/layout.jsx
@@ -1,12 +1,13 @@
-import React from 'react'
+import React, { useState } from 'react'
 import { Outlet } from "react-router-dom";
 import Header from '../../client/Header';
 import Footer from '../../client/Footer';
 
 const Layout = () => {
+  const [searchTerm, setSearchTerm] = useState("");
   return (
     <div className='god-overrall'>
-      <Header />
+      <Header searchTerm={searchTerm} setSearchTerm={setSearchTerm} />
       <Outlet />
       <Footer />
     </div>
