+++ src/config/api.handle.js
@@ -49,8 +49,8 @@ export const fetchFilmTags = () => {
 export const callCreateFilmAPI = (thumbnail, slider, name, studio, description, releaseYear, tag, category) => {
     return instance.post('/api/v1/add-film', { thumbnail, slider, name, studio, description, releaseYear, tag, category })
 }
-export const callUpdateFilmAPI = (id, thumbnail, slider, name, studio, description, releaseYear, tag, category) => {
-    return instance.put(`/api/v1/update-film`, { id, thumbnail, slider, name, studio, description, releaseYear, tag, category })
+export const callUpdateFilmAPI = (id, thumbnail, slider, name, studio, tag, category) => {
+    return instance.put(`/api/v1/update-film`, { id, thumbnail, slider, name, studio, tag, category })
 }
 export const fetchDataFilmsAPI = (query) => {
     return instance.get(`/api/v1/films?${query}`);
@@ -74,4 +74,7 @@ export const fetchSeasonsOfFilmAPI = (id, query) => {
 }
 export const AddSeasonAPI = (thumb, seasonName, description, type, releaseYear, status, filmId) => {
     return instance.post(`/api/v1/add-season`, { thumb, seasonName, description, type, releaseYear, status, film: { "id": filmId } })
+}
+export const fetchAllSeasons = (query) => {
+    return instance.get(`/api/v1/seasons?${query}`)
 }
