+++ src/pages/auth/register.page.jsx
