+++ src/pages/client/client.home.jsx
@@ -4,9 +4,11 @@ import CommingSoon from "../../components/client/CommingSoon";
 import Content from "../../components/client/Content";
 import Featured from "../../components/client/Featured";
 import { convertSlug } from "../../config/utils";
+import { useNavigate } from "react-router-dom";
 
 
 const ClientHome = () => {
+  const navigate = useNavigate();
   useEffect(() => {
     document.body.classList.add("client-theme");
     return () => {
@@ -16,8 +18,8 @@ const ClientHome = () => {
 
 
   const handleRedirect = (book) => {
-    const slug = convertSlug(book.mainText);
-    navigate(`/detail/${slug}?id=${book._id}`)
+    const slug = convertSlug(book.seasonName);
+    navigate(`/detail/${slug}?id=${book.id}`)
   }
 
   return (
